requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.2
bitarray==2.9.2
black==24.4.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
commonmark==0.9.1
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flake8==7.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20211012
pdfreader==0.1.15
pep8==1.7.1
pillow==10.3.0
pip-check-reqs==2.5.3
platformdirs==4.2.2
pre-commit==3.7.1
preshed==3.0.9
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.4
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==10.16.0
scikit-learn==1.5.0
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacypdfreader==0.3.1
srsly==2.4.8
starlette==0.37.2
thinc==8.2.5
threadpoolctl==3.5.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
